requirements.txt
google-api-python-client==1.4.1
PyYAML==3.11
urllib3==1.10.4
moviepy==0.2.2.11
youtube-dl==2015.6.25
docopt==0.6.2
